Improve memory handling through unique_ptr.

+++ src/components/ogre/BulletWorld.cpp
@@ -104,9 +104,8 @@ btCollisionWorld& BulletWorld::getCollisionWorld() const {
 // Get the mesh information for the given mesh.
 // Code found on this forum link: http://www.ogre3d.org/wiki/index.php/RetrieveVertexData
 void BulletWorld::getMeshInformation(const Ogre::MeshPtr& mesh,
-											   std::vector<float>& vertices,
-											   std::vector<unsigned int>& indices)
-{
+									 std::vector<float>& vertices,
+									 std::vector<unsigned int>& indices) {
 	bool added_shared = false;
 	size_t current_offset = 0;
 	size_t shared_offset = 0;
@@ -117,21 +116,16 @@ void BulletWorld::getMeshInformation(const Ogre::MeshPtr& mesh,
 	size_t index_count = 0;
 
 	// Calculate how many vertices and indices we're going to need
-	for (unsigned short i = 0; i < mesh->getNumSubMeshes(); ++i)
-	{
-		Ogre::SubMesh* submesh = mesh->getSubMesh( i );
+	for (size_t i = 0; i < mesh->getNumSubMeshes(); ++i) {
+		Ogre::SubMesh* submesh = mesh->getSubMesh(i);
 
 		// We only need to add the shared vertices once
-		if(submesh->useSharedVertices)
-		{
-			if( !added_shared )
-			{
+		if (submesh->useSharedVertices) {
+			if (!added_shared) {
 				vertex_count += mesh->sharedVertexData->vertexCount;
 				added_shared = true;
 			}
-		}
-		else
-		{
+		} else {
 			vertex_count += submesh->vertexData->vertexCount;
 		}
 
@@ -148,16 +142,13 @@ void BulletWorld::getMeshInformation(const Ogre::MeshPtr& mesh,
 	added_shared = false;
 
 	// Run through the submeshes again, adding the data into the arrays
-	for ( unsigned short i = 0; i < mesh->getNumSubMeshes(); ++i)
-	{
+	for (size_t i = 0; i < mesh->getNumSubMeshes(); ++i) {
 		Ogre::SubMesh* submesh = mesh->getSubMesh(i);
 
 		Ogre::VertexData* vertex_data = submesh->useSharedVertices ? mesh->sharedVertexData : submesh->vertexData;
 
-		if((!submesh->useSharedVertices)||(submesh->useSharedVertices && !added_shared))
-		{
-			if(submesh->useSharedVertices)
-			{
+		if ((!submesh->useSharedVertices) || (submesh->useSharedVertices && !added_shared)) {
+			if (submesh->useSharedVertices) {
 				added_shared = true;
 				shared_offset = current_offset;
 			}
@@ -177,8 +168,7 @@ void BulletWorld::getMeshInformation(const Ogre::MeshPtr& mesh,
 			//      Ogre::Real* pReal;
 			float* pReal;
 
-			for( size_t j = 0; j < vertex_data->vertexCount; ++j, vertex += vbuf->getVertexSize())
-			{
+			for (size_t j = 0; j < vertex_data->vertexCount; ++j, vertex += vbuf->getVertexSize()) {
 				posElem->baseVertexPointerToElement(vertex, &pReal);
 
 				vertices[(current_offset * 3) + (j * 3)] = pReal[0];
@@ -201,19 +191,14 @@ void BulletWorld::getMeshInformation(const Ogre::MeshPtr& mesh,
 		auto pShort = reinterpret_cast<Ogre::uint16*>(pLong);
 
 
-		size_t offset = (submesh->useSharedVertices)? shared_offset : current_offset;
+		size_t offset = (submesh->useSharedVertices) ? shared_offset : current_offset;
 
-		if ( use32bitindexes )
-		{
-			for ( size_t k = 0; k < numTris*3; ++k)
-			{
+		if (use32bitindexes) {
+			for (size_t k = 0; k < numTris * 3; ++k) {
 				indices[index_offset++] = static_cast<unsigned int>(pLong[k] + offset);
 			}
-		}
-		else
-		{
-			for ( size_t k = 0; k < numTris*3; ++k)
-			{
+		} else {
+			for (size_t k = 0; k < numTris * 3; ++k) {
 				indices[index_offset++] = static_cast<unsigned int>(pShort[k] + offset);
 			}
 		}
